global configuration revamp
