+++ src/components/common/GitTabs/GitProviderTabIcons.tsx
@@ -15,13 +15,14 @@
  *   limitations under the License.
  */
 
-import { GitProvider, GitProviderTabIconsProps } from './gitops.type'
-import { ReactComponent as Bitbucket } from '../../assets/icons/git/bitbucket.svg'
-import { ReactComponent as ICAwsCodeCommit } from '../../assets/icons/ic-aws-codecommit.svg'
-import { ReactComponent as GitLab } from '../../assets/icons/git/gitlab.svg'
-import { ReactComponent as GitHub } from '../../assets/icons/git/github.svg'
-import { ReactComponent as Azure } from '../../assets/icons/ic-azure.svg'
-import { ReactComponent as ICGit } from '../../assets/icons/git/git.svg'
+import { GitProviderTabIconsProps } from '../../gitOps/gitops.type'
+import { ReactComponent as Bitbucket } from '../../../assets/icons/git/bitbucket.svg'
+import { ReactComponent as ICAwsCodeCommit } from '../../../assets/icons/ic-aws-codecommit.svg'
+import { ReactComponent as GitLab } from '../../../assets/icons/git/gitlab.svg'
+import { ReactComponent as GitHub } from '../../../assets/icons/git/github.svg'
+import { ReactComponent as Azure } from '../../../assets/icons/ic-azure.svg'
+import { ReactComponent as ICGit } from '../../../assets/icons/git/git.svg'
+import { GitProvider } from './constants'
 
 const GitProviderTabIcons = ({ provider, rootClassName }: GitProviderTabIconsProps) => {
     switch (provider) {
