+++ src/components/common/GitTabs/constants.ts
@@ -1,0 +1,14 @@
+export enum GitProvider {
+    GITHUB = 'GITHUB',
+    GITLAB = 'GITLAB',
+    AWS_CODE_COMMIT = 'AWS_CODE_COMMIT',
+    AZURE_DEVOPS = 'AZURE_DEVOPS',
+    BITBUCKET_CLOUD = 'BITBUCKET_CLOUD',
+    OTHER_GIT_OPS = 'OTHER_GIT_OPS',
+}
+
+// GOOGLE = 'GOOGLE',
+// LDAP = 'LDAP',
+// MICROSOFT = 'MICROSOFT',
+// OIDC = 'OIDC',
+// OPENSHIFT = 'OPENSHIFT',
