+++ src/components/gitOps/UpdateConfirmationDialog.tsx
@@ -16,11 +16,12 @@
  */
 
 import { ButtonWithLoader, ConfirmationDialog } from '@devtron-labs/devtron-fe-common-lib'
-import GitProviderTabIcons from './GitProviderTabIcons'
-import { GitProvider, UpdateConfirmationDialogProps } from './gitops.type'
+import { getProviderNameFromEnum } from '@Components/common/GitTabs/utils'
+import { GitProvider } from '@Components/common/GitTabs/constants'
+import GitProviderTabIcons from '../common/GitTabs/GitProviderTabIcons'
+import { UpdateConfirmationDialogProps } from './gitops.type'
 import { ReactComponent as ICWarning } from '../../assets/icons/ic-warning.svg'
 import { ReactComponent as ICArrowRight } from '../../assets/icons/ic-arrow-right.svg'
-import { getProviderNameFromEnum } from './utils'
 import { DOCUMENTATION } from '../../config'
 
 const UpdateConfirmationDialog = ({
