+++ src/components/gitOps/gitops.type.ts
@@ -17,6 +17,7 @@
 import { RouteComponentProps } from 'react-router'
 import { TLSConfigDTO, TLSConnectionDTO } from '../common/TLSConnectionForm/types'
 import { BaseGitOpsType, GitOpsAuthModeType } from '@devtron-labs/devtron-fe-common-lib'
+import { GitProvider } from '@Components/common/GitTabs/constants'
 
 export type GitOpsOrganisationIdType =
     | 'gitHubOrgId'
@@ -25,14 +26,14 @@ export type GitOpsOrganisationIdType =
     | 'bitBucketWorkspaceId'
     | 'bitBucketProjectKey'
 
-export enum GitProvider {
-    GITHUB = 'GITHUB',
-    GITLAB = 'GITLAB',
-    AWS_CODE_COMMIT = 'AWS_CODE_COMMIT',
-    AZURE_DEVOPS = 'AZURE_DEVOPS',
-    BITBUCKET_CLOUD = 'BITBUCKET_CLOUD',
-    OTHER_GIT_OPS = 'OTHER_GIT_OPS',
-}
+// export enum GitProvider {
+//     GITHUB = 'GITHUB',
+//     GITLAB = 'GITLAB',
+//     AWS_CODE_COMMIT = 'AWS_CODE_COMMIT',
+//     AZURE_DEVOPS = 'AZURE_DEVOPS',
+//     BITBUCKET_CLOUD = 'BITBUCKET_CLOUD',
+//     OTHER_GIT_OPS = 'OTHER_GIT_OPS',
+// }
 
 export type GitProviderType = GitProvider | 'BITBUCKET_DC'
